# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PPTX Voiceover Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Slide {
  index: number;
  hasAudio: boolean;
  /** @nullable */
  audioDurationSeconds?: number | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Presentation {
  id: string;
  filename: string;
  slideCount: number;
  slides: Slide[];
  createdAt: string;
  /** @nullable */
  exportStatus?: string | null;
}

export type ExportJobStatus = typeof ExportJobStatus[keyof typeof ExportJobStatus];


export const ExportJobStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  error: 'error',
} as const;

export interface ExportJob {
  presentationId: string;
  status: ExportJobStatus;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  downloadUrl?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface ErrorResponse {
  error: string;
}