# lib/api-zod/src/generated/types/exportJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PPTX Voiceover Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ExportJobStatus } from './exportJobStatus';

export interface ExportJob {
  presentationId: string;
  status: ExportJobStatus;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  downloadUrl?: string | null;
  /** @nullable */
  error?: string | null;
}