# lib/api-zod/src/generated/types/presentation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PPTX Voiceover Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { Slide } from './slide';

export interface Presentation {
  id: string;
  filename: string;
  slideCount: number;
  slides: Slide[];
  createdAt: string;
  /** @nullable */
  exportStatus?: string | null;
}