# lib/api-zod/src/generated/types/slide.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PPTX Voiceover Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Slide {
  index: number;
  hasAudio: boolean;
  /** @nullable */
  audioDurationSeconds?: number | null;
  /** @nullable */
  imageUrl?: string | null;
}